ADD Changes in AUTH permissions

+++ clientes/views.py
@@ -1,4 +1,6 @@
-from django.shortcuts import render,redirect,HttpResponse
+from django.shortcuts import HttpResponse
+from django.shortcuts import render
+from django.shortcuts import redirect
 from django.shortcuts import get_object_or_404
 from .models import Person
 from vendas.models import Venda
@@ -11,50 +13,101 @@ from django.views.generic.edit import CreateView
 from django.views.generic.edit import UpdateView,DeleteView
 from django.urls import reverse_lazy
 from django.views import View
+from django.utils.decorators import method_decorator
+
 
 
 
 @login_required
 def person_list(request):
-    termo_busca = request.GET.get('pesquisa' ,  None)
 
-    if termo_busca:
-        persons = Person.objects.all()
-        persons = persons.filter(first_name__icontains=termo_busca)
+    if request.user.has_perm('clientes.view_person') == True:
+
+        termo_busca = request.GET.get('pesquisa' ,  None)
+
+        usertemp = request.user
+
+        if usertemp.is_staff == True:
+
+            if termo_busca:
+                persons = Person.objects.all()
+                persons = persons.filter(first_name__icontains=termo_busca)
+            else:
+                persons = Person.objects.all()
+
+            return render(request,'person.html',{"persons":persons})
+        else:
+            return render(request, "403.html")
+
     else:
-        persons = Person.objects.all()
-
-    return render(request,'person.html',{"persons":persons})
+        return render(request,"403.html")
 
 @login_required
 def person_new(request):
-    form = PersonForm(request.POST or None,request.FILES or None)
 
-    if form.is_valid():
-        form.save()
-        return redirect('person_list')
+    if request.user.has_perm("clientes.add_person") == True:
 
-    return render(request,"person_form.html",{"form":form})
+        usertemp = request.user
+
+        if usertemp.is_staff == True:
+
+            form = PersonForm(request.POST or None,request.FILES or None)
+
+            if form.is_valid():
+                form.save()
+                return redirect('person_list')
+
+            return render(request,"person_form.html",{"form":form})
+        else:
+            return render(request,"403.html")
+    else:
+        return render(request, "403.html")
 
 @login_required
 def person_update(request,id):
-    person   = get_object_or_404(Person,pk=id)
-    form = PersonForm(request.POST or None, request.FILES or None,instance=person)
 
-    if form.is_valid():
-        form.save()
-        return redirect('person_list')
 
-    return render(request,"person_form.html",{"form":form})
+    if request.user.has_perm("clientes.update_person") == True:
+
+        usertemp = request.user
+
+        if usertemp.is_staff == True:
+
+            person   = get_object_or_404(Person,pk=id)
+            form = PersonForm(request.POST or None, request.FILES or None,instance=person)
+
+            if form.is_valid():
+                form.save()
+                return redirect('person_list')
+
+            return render(request,"person_form.html",{"form":form})
+        else:
+            return render(request,"403.html")
+    else:
+        return render(request,"403.html")
+
 
 @login_required
 def person_delete(request,id):
-    person = get_object_or_404(Person,pk=id)
 
-    if request.method == "POST":
-        person.delete()
-        return redirect('person_list')
-    return  render(request,"person_delete_confirm.html", {'person':person})
+    if request.user.has_perm("clientes.delete_person") == True:
+        usertemp = request.user
+
+        if usertemp.is_staff == True:
+            person = get_object_or_404(Person, pk=id)
+
+            if request.method == "POST":
+                person.delete()
+                return redirect('person_list')
+            return render(request, "person_delete_confirm.html", {'person': person})
+
+        else:
+            return render(request,"403.html")
+    else:
+        return render(request,"403.html")
+
+
+
 
 #Cbv Person
 
@@ -65,7 +118,6 @@ class PersonList(ListView):
 
     def  get_queryset(self):
         termo_busca = self.request.GET.get('pesquisa' , None)
-
 
         if termo_busca:
             persons = Person.objects.filter(first_name__icontains=termo_busca)
@@ -81,6 +133,8 @@ class PersonDetail(DetailView):
     model = Person
 
 
+
+
     def get_object(self, queryset=None):
         pk = self.kwargs.get(self.pk_url_kwarg)
         return Person.objects.select_related('doc').get(id=pk)
@@ -91,6 +145,7 @@ class PersonDetail(DetailView):
             pessoa_id=self.object.id
         )
         return context
+
 
 
 
@@ -106,8 +161,6 @@ class PersonUpdate(UpdateView):
     model = Person
     fields = ['first_name','last_name','age','salary','bio','photo','doc']
     #success_url = reverse_lazy('person_list_cbv')
-
-
 
     def get_success_url(self):
         return  reverse_lazy('person_list_cbv')
@@ -132,5 +185,9 @@ class ProdutoBulk(View):
 
         #Produto.objects.bulk_create(list_produtos)
 
-
         return HttpResponse('Funcionou')
+
+
+
+
+
